Simplify using hook
